* commit 'e55a04c1915c62bbc23ee1e5db70208ce2a11679': (403 commits)
  Support sa tokens for /metrics path
  reset metrics before reporting
  Enable feature version support for KDM releases (#36558)
  fix healthcheck serialization
  don't error if periodic instruction output is not found (#36566)
  fix create_db parameter resource
  Fix typo in rancher-tiller
  go generate
  Migrate sourcecodeproviderconfig creds to secrets
  Migrate catalog passwords to Secrets
  Migrate secrets from notifiers
  Sanitize cluster template questions and answers
  Migrate Weave CNI password to Secret
  Migrate S3 backup credentials to Secrets
  Migrate cluster registry passwords to Secrets
  Print base64 kubeconfig for k3s and rke2 clusters
  add wins install script for rancher airgap (#36553)
  use more effective shutdown command for certificate rotation
  Fixing various bugs with the catalog role bindings
  Upgrading version of loglevel
  ...

# Conflicts:
#	Dockerfile.dapper
#	package/Dockerfile

+++ package/jailer.sh
@@ -58,13 +58,17 @@ cp -r -l /opt/drivers/management-state/bin /opt/jail/$NAME/var/lib/rancher/manag
 cp -l /usr/bin/rancher-machine /opt/jail/$NAME/usr/bin
 
 # Hard link helm_2 into the jail
-cp -l /usr/bin/rancher-helm /opt/jail/$NAME/usr/bin
+if [[ -f /usr/bin/rancher-helm ]]; then
+  cp -l /usr/bin/rancher-helm /opt/jail/$NAME/usr/bin
+fi
 
 # Hard link helm_3 into the jail
 cp -l /usr/bin/helm_v3 /opt/jail/$NAME/usr/bin
 
 # Hard link tiller into the jail
-cp -l /usr/bin/rancher-tiller /opt/jail/$NAME/usr/bin
+if [[ -f /usr/bin/rancher-tiller ]]; then
+  cp -l /usr/bin/rancher-tiller /opt/jail/$NAME/usr/bin
+fi
 
 # Hard link kustomize into the jail
 cp -l /usr/bin/kustomize /opt/jail/$NAME/usr/bin
@@ -75,11 +79,17 @@ cp -l /usr/bin/kustomize.sh /opt/jail/$NAME
 # Hard link ssh into the jail
 cp -l /usr/bin/ssh /opt/jail/$NAME/usr/bin
 
+# Hard link nc into the jail
+cp -l /usr/bin/nc /opt/jail/$NAME/usr/bin
+
 # Hard link cat into the jail
 cp -l /bin/cat /opt/jail/$NAME/bin/
 
 # Hard link bash into the jail
 cp -l /bin/bash /opt/jail/$NAME/bin/
+
+# Hard link sh into the jail
+cp -l /bin/sh /opt/jail/$NAME/bin/
 
 # Hard link rm into the jail
 cp -l /bin/rm /opt/jail/$NAME/bin/
